// apps/console/src/pages/organizations/policies/__generated__/PolicyViewQuery.graphql.ts
export type PolicyStatus = "ACTIVE" | "DRAFT";
export type PolicyViewQuery$variables = {
  policyId: string;
};
export type PolicyViewQuery$data = {
  readonly node: {
    readonly content?: string;
    readonly createdAt?: string;
    readonly id: string;
    readonly name?: string;
    readonly owner?: {
      readonly fullName: string;
      readonly id: string;
      readonly primaryEmailAddress: string;
    };
    readonly reviewDate?: string | null | undefined;
    readonly status?: PolicyStatus;
    readonly updatedAt?: string;
  };
};
export type PolicyViewQuery = {
  response: PolicyViewQuery$data;
  variables: PolicyViewQuery$variables;
};